// order-ticket-load-core/src/domain/enums/delivery-state.ts
export enum DeliveryState {
  PLANT_DEPARTED = 'PLANT_DEPARTED',
  GEOFENCE_ENTERED = 'GEOFENCE_ENTERED',
  ON_SITE_ARRIVED = 'ON_SITE_ARRIVED',
  POUR_STARTED = 'POUR_STARTED',
  POUR_COMPLETED = 'POUR_COMPLETED',
  SITE_DEPARTED = 'SITE_DEPARTED',
  PLANT_RETURNED = 'PLANT_RETURNED',
  WASHOUT_COMPLETED = 'WASHOUT_COMPLETED',
}